Adafuitskript angepasst für Python 3.6

+++ Adafruit/Adafruit_BMP180.py
@@ -1,11 +1,13 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*
 import time
-from Adafruit_I2C import Adafruit_I2C
+from Adafruit.Adafruit_I2C import Adafruit_I2C
 
 # ===========================================================================
 # BMP085 Class
-# Anpassung des  Pint Befehls für Python 3.6
+# Anpassung des  Print Befehls für Python 3.6
+# Anpassung int() >>
+# kein Anspruch auf Richtigkeit der Änderungen
 # ===========================================================================
 
 class BMP085 :
@@ -129,7 +131,7 @@ class BMP085 :
     msb = self.i2c.readU8(self.__BMP085_PRESSUREDATA)
     lsb = self.i2c.readU8(self.__BMP085_PRESSUREDATA+1)
     xlsb = self.i2c.readU8(self.__BMP085_PRESSUREDATA+2)
-    raw = ((msb << 16) + (lsb << 8) + xlsb) >> (8 - self.mode)
+    raw =int((msb << 16) + (lsb << 8) + xlsb) >> (8 - self.mode)
     if (self.debug):
       print ("DBG: Raw Pressure: 0x%04X (%d)" % (raw & 0xFFFF, raw))
     return raw
@@ -144,10 +146,10 @@ class BMP085 :
 
     # Read raw temp before aligning it with the calibration values
     UT = self.readRawTemp()
-    X1 = ((UT - self._cal_AC6) * self._cal_AC5) >> 15
-    X2 = (self._cal_MC << 11) / (X1 + self._cal_MD)
+    X1 = int((UT - self._cal_AC6) * self._cal_AC5) >> 15
+    X2 = (int(self._cal_MC) << 11) / (X1 + self._cal_MD)
     B5 = X1 + X2
-    temp = ((B5 + 8) >> 4) / 10.0
+    temp = (int(B5 + 8) >> 4) / 10.0
     if (self.debug):
       print ("DBG: Calibrated temperature = %f C" % temp)
     return temp
@@ -192,19 +194,19 @@ class BMP085 :
         self.showCalibrationData()
 
     # True Temperature Calculations
-    X1 = ((UT - self._cal_AC6) * self._cal_AC5) >> 15
-    X2 = (self._cal_MC << 11) / (X1 + self._cal_MD)
+    X1 = int((UT - self._cal_AC6) * self._cal_AC5) >> 15
+    X2 = (int(self._cal_MC) << 11) / (X1 + self._cal_MD)
     B5 = X1 + X2
     if (self.debug):
       print ("DBG: X1 = %d" % (X1))
       print ("DBG: X2 = %d" % (X2))
       print ("DBG: B5 = %d" % (B5))
-      print ("DBG: True Temperature = %.2f C" % (((B5 + 8) >> 4) / 10.0))
+      print ("DBG: True Temperature = %.2f C" % ((int(B5 + 8) >> 4) / 10.0))
 
     # Pressure Calculations
     B6 = B5 - 4000
-    X1 = (self._cal_B2 * (B6 * B6) >> 12) >> 11
-    X2 = (self._cal_AC2 * B6) >> 11
+    X1 = int(self._cal_B2 *int(B6 * B6) >> 12) >> 11
+    X2 = int(self._cal_AC2 * B6) >> 11
     X3 = X1 + X2
     B3 = (((self._cal_AC1 * 4 + X3) << self.mode) + 2) / 4
     if (self.debug):
@@ -214,10 +216,10 @@ class BMP085 :
       print ("DBG: X3 = %d" % (X3))
       print ("DBG: B3 = %d" % (B3))
 
-    X1 = (self._cal_AC3 * B6) >> 13
-    X2 = (self._cal_B1 * ((B6 * B6) >> 12)) >> 16
-    X3 = ((X1 + X2) + 2) >> 2
-    B4 = (self._cal_AC4 * (X3 + 32768)) >> 15
+    X1 = int(self._cal_AC3 * B6) >> 13
+    X2 = int(self._cal_B1 * (int(B6 * B6) >> 12)) >> 16
+    X3 = int((X1 + X2) + 2) >> 2
+    B4 = int(self._cal_AC4 * (X3 + 32768)) >> 15
     B7 = (UP - B3) * (50000 >> self.mode)
     if (self.debug):
       print ("DBG: X1 = %d" % (X1))
@@ -234,15 +236,15 @@ class BMP085 :
     if (self.debug):
       print ("DBG: X1 = %d" % (X1))
       
-    X1 = (p >> 8) * (p >> 8)
-    X1 = (X1 * 3038) >> 16
-    X2 = (-7357 * p) >> 16
+    X1 = (int(p) >> 8) * (int(p) >> 8)
+    X1 = int(X1 * 3038) >> 16
+    X2 = int(-7357 * p) >> 16
     if (self.debug):
       print ("DBG: p  = %d" % (p))
       print ("DBG: X1 = %d" % (X1))
       print ("DBG: X2 = %d" % (X2))
 
-    p = p + ((X1 + X2 + 3791) >> 4)
+    p = p + (int(X1 + X2 + 3791) >> 4)
     if (self.debug):
       print ("DBG: Pressure = %d Pa" % (p))
